fix(workflow): isolate import helpers (#12)
